requirements.txt
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
msgpack==1.1.0
narwhals==1.14.1
numba==0.60.0
numpy==1.26.4
openai==1.54.5
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.7
pooch==1.8.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
pyworld==0.3.5
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.7.4
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.41.3
threadpoolctl==3.6.0
tomlkit==0.12.0
tqdm==4.67.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==11.0.3
zipp==3.21.0
gradio>=4.44.1
torch
pyyaml
